scripts/requirements.txt
PyGithub>=2.4.0
